Move rsspp::Parser into a separate file
